v3.23.1.7 Função de cadastrar carro

+++ Automoveis/Automovel.java
@@ -5,7 +5,7 @@ public abstract class Automovel extends Formulario.FormRefactored {
     private String marca;
     private String modelo;
     private int idade;
-    private double kilomt;
+    private long kilomt;
     private String tipo;
     private String combustivel;
     private int qt_marcha;
@@ -14,7 +14,7 @@ public abstract class Automovel extends Formulario.FormRefactored {
     private int tamanhotanque;
     private double valor;
 
-    public Automovel(String marca, String modelo, int idade, double kilomt, String tipo, String combustivel, int qt_marcha, String cor, String chassi, int tamanhotanque, double valor) throws HeadlessException {
+    public Automovel(String marca, String modelo, int idade, long kilomt, String tipo, String combustivel, int qt_marcha, String cor, String chassi, int tamanhotanque, double valor) throws HeadlessException {
         this.marca = marca;
         this.modelo = modelo;
         this.idade = idade;
